Hack to get around uwsgi reload issue
Sometimes, when there's a couple of uwsgi reloads in quick succession and you
have inbound requests, some requests can time out. I suspect this is because
threading.Event().wait() can't be interrupted, and this can interfere with uwsgi
trying to interrupt the process. This change bypasses threading.Event() and uses a
sleep.